+++ src/transport/mmio.rs
@@ -442,7 +442,7 @@ impl Transport for MmioTransport {
         }
     }
 
-    fn config_space(&self) -> NonNull<u64> {
+    fn config_space(&self) -> NonNull<u32> {
         NonNull::new((self.header.as_ptr() as usize + CONFIG_SPACE_OFFSET) as _).unwrap()
     }
 }
